+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## version 0.2.1+ (Unreleased)
+- add `inshour` subcommand
 
 ## version 0.2.1 (2018-02-27)
 - add --version option